requirements.txt
Flask
supabase
pandas
beautifulsoup4
playwright
python-dotenv
